A somewhat working prototype?
Right now, the mod erases chunks based on the `InhabitedTime` property. Whether that is a good idea or not I do not know, needs testing. The game also for some reason fills in deleted chunks with empty data, causing the region files to maintain their size, regardless of how much data is removed. This needs to be worked around somehow.